Implement torch.diagonal for named tensors
Featuring:
- Added a NoNamesGuard::reset() function that sets NamesMode back to
what it was before the guard. This makes it so that we don't have to
create a new context to run code in an unnamed way.
- Added a diagonal(Tensor, *, Dimname outdim, Dimname dim1, Dimname dim2, int64_t offset=0)
overload. All of the non-tensor arguments are keyword only for
readability purposes; something like `tensor.diagonal("A", "B", "C")`
would be really confusing.

Test Plan:
- Added new tests

ghstack-source-id: 4f35ef5e913b7c9af74bcaa6e8c6cb49a4d5765d
Pull Request resolved: https://github.com/pytorch/pytorch/pull/30193